refactor: move helper files out of app/
