+++ packages/web/app/beta/layout.tsx
@@ -1,7 +1,7 @@
 "use strict";
 
-import Footer from "../components/Footer";
-import Navbar from "../components/Navbar";
+import Footer from "../../components/Footer";
+import Navbar from "../../components/Navbar";
 
 export default function RootLayout({
   children,
