+++ packages/web/components/Footer.tsx
