+++ packages/web/components/HomeBanner.tsx
