+++ packages/web/components/QueryResult.tsx
